requirements.txt
Flask
stripe
python-docx
requests
openai-whisper
